trained baseline fencing model

+++ phc/env/util/gym_util.py
@@ -185,16 +185,16 @@ def get_robot_states_from_torch_tensor(config, ts, global_quats, vels, avels,
 
     if config['env_frame_ob']:
         if type(motion_length) == np.ndarray:
-            motion_length = motion_length.astype(np.float)
-            progress_ob = np.expand_dims(progress.astype(np.float) /
+            motion_length = motion_length.astype(np.float64)
+            progress_ob = np.expand_dims(progress.astype(np.float64) /
                                          motion_length, axis=-1)
         else:
-            progress_ob = np.expand_dims(progress.astype(np.float) /
+            progress_ob = np.expand_dims(progress.astype(np.float64) /
                                          float(motion_length), axis=-1)
         obs = np.concatenate([obs, progress_ob], axis=-1)
 
     if config['env_motion_ob'] and not config['env_motion_ob_onehot']:
-        motion_id_ob = np.expand_dims(motion_id.astype(np.float) /
+        motion_id_ob = np.expand_dims(motion_id.astype(np.float64) /
                                       float(num_motion), axis=-1)
         obs = np.concatenate([obs, motion_id_ob], axis=-1)
     elif config['env_motion_ob'] and config['env_motion_ob_onehot']:
